// job_board_frontend_project/interfaces/Recruiter.ts
// interfaces/Recruiter.ts

export interface RecruiterForm {
  username: string;
  email: string;
  first_name: string;
  last_name: string;
  password: string;
  company_name: string;
  company_size: string;
  industry: string;
  country: string;
  state: string;
  phone_number: string;
  website: string;
  description: string;
}